Move the “wasm stack” to the bottom of memory
Move the wasm stack to the bottom of the linear memory. This has the benefit of causing a page fault when the stack overruns (due to wrap around to 0xffffffff which is always a fault in our environment) vs the current behavior of overrunning in to heap and other variables which tends to cause difficult to understand problems.

Example layout difference of the eosio.system contract

Previously:
0KB - 8KB; SP; strings, globals, etc
8KB - 16KB; first 8KB of heap
16KB - 26KB; 10KB of stack
26KB - 64KB; unused memory

With this commit:
0KB-10KB; 10KB of stack
10KB - 18KB; SP; strings; globals, etc
18KB - 26KB; first 8KB of heap
26KB - 64KB; unused memory